Disentangle Modifies from HyperStack
Previously, FStar.HyperStack used to depend on
FStar.Modifies, so that `make -C ulib mgen` had to
extract FStar.Modifies, which failed due to a bug
similar to #490.

This commit breaks this dependency by creating a
new FStar.HyperStackNG module that is specific to the
new library. (In particular, this also shows that the
theory of general `modifies` clauses can be
instantiated to HyperStack without breaking any
abstractions.)