Final level-0 objects; simplify modifies if tip preserved
An object of class level 0 is final only if it cannot
be inherited (i.e. no object of higher-level class can
be built on top of it), and it is disjoint from any
other object of the same class.

This support is necessary to prove that if
(tip \cup S) is modified but the tip is preserved,
then S is modified.